+++ src/test/java/com/backend/immilog/user/application/UserInformationServiceTest.java
@@ -1,15 +1,12 @@
 package com.backend.immilog.user.application;
 
 import com.backend.immilog.global.application.ImageService;
-import com.backend.immilog.global.application.RedisService;
 import com.backend.immilog.global.exception.CustomException;
-import com.backend.immilog.global.security.JwtProvider;
 import com.backend.immilog.user.application.command.UserPasswordChangeCommand;
 import com.backend.immilog.user.application.services.UserInformationServiceImpl;
-import com.backend.immilog.user.model.enums.UserStatus;
-import com.backend.immilog.user.model.dtos.UserSignInDTO;
 import com.backend.immilog.user.model.embeddables.Location;
 import com.backend.immilog.user.model.entities.User;
+import com.backend.immilog.user.model.enums.UserStatus;
 import com.backend.immilog.user.model.repositories.UserRepository;
 import com.backend.immilog.user.model.services.UserInformationService;
 import com.backend.immilog.user.presentation.request.UserInfoUpdateRequest;
@@ -25,25 +22,21 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import java.util.Optional;
 import java.util.concurrent.CompletableFuture;
 
-import static com.backend.immilog.user.model.enums.Countries.*;
-import static com.backend.immilog.user.model.enums.UserRole.ROLE_ADMIN;
-import static com.backend.immilog.user.model.enums.UserRole.ROLE_USER;
+import static com.backend.immilog.global.enums.Countries.*;
+import static com.backend.immilog.global.enums.UserRole.ROLE_ADMIN;
+import static com.backend.immilog.global.enums.UserRole.ROLE_USER;
 import static com.backend.immilog.user.exception.UserErrorCode.NOT_AN_ADMIN_USER;
 import static com.backend.immilog.user.exception.UserErrorCode.PASSWORD_NOT_MATCH;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.mockito.Mockito.*;
+import static org.mockito.Mockito.when;
 
 @DisplayName("사용자 정보 서비스 테스트")
 class UserInformationServiceTest {
     @Mock
     private UserRepository userRepository;
     @Mock
-    private JwtProvider jwtProvider;
-    @Mock
     private PasswordEncoder passwordEncoder;
-    @Mock
-    private RedisService redisService;
     @Mock
     private ImageService imageService;
     private UserInformationService userInformationService;
@@ -53,74 +46,9 @@ class UserInformationServiceTest {
         MockitoAnnotations.openMocks(this);
         userInformationService = new UserInformationServiceImpl(
                 userRepository,
-                jwtProvider,
                 passwordEncoder,
-                redisService,
                 imageService
         );
-    }
-
-    @Test
-    @DisplayName("사용자 정보 조회 성공 : 위치 정보 일치")
-    void getUserSignInDTO_success() {
-        // given
-        Long userSeq = 1L;
-        User user = User.builder()
-                .seq(userSeq)
-                .email("test@email.com")
-                .nickName("test")
-                .imageUrl("image")
-                .userStatus(UserStatus.PENDING)
-                .userRole(ROLE_USER)
-                .interestCountry(SOUTH_KOREA)
-                .location(Location.of(SOUTH_KOREA, "Seoul"))
-                .reportInfo(null)
-                .build();
-
-        Pair<String, String> country = Pair.of("South Korea", "Seoul");
-        when(userRepository.findById(userSeq)).thenReturn(Optional.of(user));
-        when(jwtProvider.issueAccessToken(any())).thenReturn("accessToken");
-        when(jwtProvider.issueRefreshToken()).thenReturn("refreshToken");
-
-        // when
-        UserSignInDTO result = userInformationService.getUserSignInDTO(userSeq,
-                country);
-        // then
-        verify(redisService, times(1)).saveKeyAndValue(
-                "Refresh: refreshToken", user.getEmail(), 5 * 29 * 24 * 60
-        );
-        assertThat(result.userSeq()).isEqualTo(userSeq);
-    }
-
-    @Test
-    @DisplayName("사용자 정보 조회 성공 : 위치 정보 불 일치")
-    void getUserSignInDTO_success_location_not_match() {
-        // given
-        Long userSeq = 1L;
-        User user = User.builder()
-                .seq(userSeq)
-                .email("test@email.com")
-                .nickName("test")
-                .imageUrl("image")
-                .userStatus(UserStatus.PENDING)
-                .userRole(ROLE_USER)
-                .interestCountry(SOUTH_KOREA)
-                .location(Location.of(MALAYSIA, "KL"))
-                .reportInfo(null)
-                .build();
-
-        Pair<String, String> country = Pair.of("South Korea", "Seoul");
-        when(userRepository.findById(userSeq)).thenReturn(Optional.of(user));
-        when(jwtProvider.issueAccessToken(any())).thenReturn("accessToken");
-        when(jwtProvider.issueRefreshToken()).thenReturn("refreshToken");
-
-        // when
-        UserSignInDTO result = userInformationService.getUserSignInDTO(userSeq, country);
-        // then
-        verify(redisService, times(1)).saveKeyAndValue(
-                "Refresh: refreshToken", user.getEmail(), 5 * 29 * 24 * 60
-        );
-        assertThat(result.userSeq()).isEqualTo(userSeq);
     }
 
     @Test
