+++ src/test/java/com/backend/immilog/user/application/UserReportServiceTest.java
@@ -4,11 +4,11 @@ import com.backend.immilog.global.application.RedisDistributedLock;
 import com.backend.immilog.global.exception.CustomException;
 import com.backend.immilog.user.application.services.UserReportServiceImpl;
 import com.backend.immilog.user.enums.ReportReason;
-import com.backend.immilog.user.model.enums.UserStatus;
 import com.backend.immilog.user.model.embeddables.Location;
 import com.backend.immilog.user.model.embeddables.ReportInfo;
 import com.backend.immilog.user.model.entities.Report;
 import com.backend.immilog.user.model.entities.User;
+import com.backend.immilog.user.model.enums.UserStatus;
 import com.backend.immilog.user.model.repositories.ReportRepository;
 import com.backend.immilog.user.model.repositories.UserRepository;
 import com.backend.immilog.user.model.services.UserReportService;
@@ -23,9 +23,9 @@ import java.sql.Date;
 import java.time.LocalDateTime;
 import java.util.Optional;
 
-import static com.backend.immilog.user.model.enums.Countries.MALAYSIA;
-import static com.backend.immilog.user.model.enums.Countries.SOUTH_KOREA;
-import static com.backend.immilog.user.model.enums.UserRole.ROLE_USER;
+import static com.backend.immilog.global.enums.Countries.MALAYSIA;
+import static com.backend.immilog.global.enums.Countries.SOUTH_KOREA;
+import static com.backend.immilog.global.enums.UserRole.ROLE_USER;
 import static com.backend.immilog.user.exception.UserErrorCode.ALREADY_REPORTED;
 import static com.backend.immilog.user.exception.UserErrorCode.CANNOT_REPORT_MYSELF;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
